Form check and bug fixing

+++ Progetto-TIW/src/main/java/it/polimi/tiw/dao/ItemDAO.java
@@ -76,7 +76,7 @@ public class ItemDAO {
 			 */
 			rs = ps.executeQuery();
 			if (rs.next())
-				out = rs.getInt("tot");
+				out = rs.getFloat("tot");
 
 		} catch (SQLException e) {
 			throw e;
